Refactor the refactor
