+++ src/ILCompiler.Compiler/src/Compiler/TypeExtensions.cs
@@ -194,5 +194,32 @@ namespace ILCompiler
             TypeDesc elementType = arrayType.ElementType;
             return type.IsMdArray || elementType.IsPointer || elementType.IsFunctionPointer;
         }
+
+        /// <summary>
+        /// Determines whether an object of type '<paramref name="type"/>' requires 8-byte alignment on 
+        /// 32bit ARM architectures.
+        /// </summary>
+        public static bool RequiresAlign8(this TypeDesc type)
+        {
+            if (type.Context.Target.Architecture != TargetArchitecture.ARM)
+            {
+                return false;
+            }
+
+            if (type.IsArray)
+            {
+                var elementType = ((ArrayType)type).ElementType;
+                if ((elementType.IsValueType) && ((DefType)elementType).InstanceByteAlignment.AsInt > 4)
+                {
+                    return true;
+                }
+            }
+            else if (type.IsDefType && ((DefType)type).InstanceByteAlignment.AsInt > 4)
+            {
+                return true;
+            }
+
+            return false;
+        }
     }
 }
