+++ src/JitInterface/src/MemoryHelper.cs
@@ -3,12 +3,8 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace ILCompiler
+namespace Internal.JitInterface
 {
     internal static unsafe class MemoryHelper
     {
